mandatory all done

+++ stack_functions/ft_stacksize.c
@@ -6,15 +6,15 @@
 /*   By: ltruchel <ltruchel@student.42angoulem      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/04 12:32:20 by ltruchel          #+#    #+#             */
-/*   Updated: 2022/12/03 20:21:33 by ltruchel         ###   ########.fr       */
+/*   Updated: 2022/12/07 18:51:50 by ltruchel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
-unsigned int	ft_stacksize(t_stack *stack)
+size_t	ft_stacksize(t_stack *stack)
 {
-	unsigned int	i;
+	size_t	i;
 
 	if (!stack)
 		return (0);
